Deploy to static site to gh-pages
Use a GitHub actions to build the website and publish it to the gh-pages
branch.

Inspired from
https://github.com/marketplace/actions/github-pages-action#%EF%B8%8F-static-site-generators-with-nodejs

The GitHub action can handle the CNAME, remove the drop-in files from
source.

The project layout was updated to match the default Vite project layout:

- The content of src/ was moved back to the top level.
- The build process outputs to dist/, and is not tracked by git. The
  docs directory was removed.

That allowed simplifying the README and configuration considerably.